+++ src/pat/contentbrowser/src/ContentBrowser.svelte
@@ -5,7 +5,7 @@
     import { fly } from "svelte/transition";
     import _t from "../../../core/i18n-wrapper";
     import { ensureIntlSupport } from "../../../core/intl-loader";
-    import Upload from "../../upload/upload";
+    import Upload from "../../upload/upload--implementation";
     import contentStore from "./ContentStore";
     import {
         clickOutside,
