+++ src/pat/structure/js/views/upload.js
@@ -1,7 +1,7 @@
 import $ from "jquery";
 import _ from "underscore";
 import PopoverView from "../../../../core/ui/views/popover";
-import Upload from "../../../upload/upload";
+import Upload from "../../../upload/upload--implementation";
 
 export default PopoverView.extend({
     className: "popover upload",
